Adding C++ unit tests to hit contrib directory
